requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anomalib==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_43o9lm92p9/croot/certifi_1750863864870/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
docstring_parser==0.16
einops==0.8.0
et_xmlfile==2.0.0
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools==4.58.5
FrEIA==0.2
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.33.2
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.37.0
imgaug==0.4.0
importlib_resources==6.5.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.5.1
jsonargparse==4.34.0
kiwisolver==1.4.8
kornia==0.6.9
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.9
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.6.3
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.32.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.3.1
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.11.6
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich==14.0.0
rich-argparse==1.7.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.6.12
torch==2.4.1
torchmetrics==1.7.4
torchvision==0.19.1
tqdm==4.67.1
typeshed_client==2.7.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.8
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wcwidth==0.2.13
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
yarl==1.20.1
